Refactor `install_driver/tasks/ubuntu.yaml`
When provisioning a new driver `450` machine recently, I ran into an error that said an incompatible version of `libnvidia-ifr1` was going to be installed.

Adding `libnvidia-ifr1` to the existing list of dependencies resolved this.

However, I then proceeded to provision a driver `495` machine and received an error message stating that the `libnvidia-ifr1` package wasn't available for driver `495`.

To account for the discrepancy in packages between driver `450` and `495` machines, I've refactored `install_driver/tasks` to include:

- a list of common packages that should be used for all driver versions
- a separate list of packages (containing only `libnvidia-ifr1` at this time) that should be conditionally appended to the common packages
  - the condition for when `libnvidia-ifr1` should be appended was determined to be `470>= driver_version >=450`