Use gcc10 to build kernel modules on Amazon Linux (#2)
`gcc7` is the default version on CentOS 7, but the latest `nvidia-open`
kernel modules requires `gcc10` to build, so we install `gcc10` and set
it as default before installing the driver packages. `yum` will install
the packages and build the kernel modules using `dkms` which will use
`gcc10` to build. We also restore `gcc7` as default version once the
kernel modules are built.

Packer PR to test these changes:
https://github.com/rapidsai/packer-templates/pull/6

Signed-off-by: Jordan Jacobelli <jjacobelli@nvidia.com>